requirements.txt
asgiref==3.3.4
attrs==21.4.0
boto3==1.17.94
botocore==1.20.94
coverage==6.3.2
Django==3.1.4
django-environ==0.4.5
django-extensions==3.1.0
djangorestframework==3.12.2
environ==1.0
execnet==1.9.0
iniconfig==1.1.1
jmespath==0.10.0
mysql-connector-python==8.0.23
packaging==21.3
pluggy==1.0.0
protobuf==3.17.3
py==1.11.0
pycryptodomex==3.10.1
PyJWT==2.1.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-dateutil==2.8.1
pytz==2021.1
redis==3.5.3
s3transfer==0.4.2
six==1.16.0
sqlparse==0.4.1
tomli==2.0.1
urllib3==1.26.5
Werkzeug==2.0.1
